requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
black==22.8.0
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.50.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-leaflet==0.28.2
django-prometheus==2.3.1
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.2
drf-yasg==1.21.4
feedgen==0.9.0
flake8==6.0.0
hyperlink==21.0.0
idna==3.3
importlib-resources==5.10.0
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
lxml==4.9.3
MarkupSafe==2.1.2
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==23.0
pathspec==0.10.1
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
prometheus-client==0.17.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.15.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.9.0
zope.interface==5.5.2
